# concurrency,kvserver: limited scans do optimistic latching